Update PYTHONPATH in Docker workflow
The PYTHONPATH used in the .github/workflows/docker-image.yml file has been updated. It was previously set to 'api-automation-framework', but now it points directly to the 'src' directory in the present working directory. This change ensures that Python modules are correctly found and imported during the Docker image build process.